refactor: Separate `*Expr`, `*Scalar`
Gonna be soooooooo much easier to work on them side-by-side now 🥳
