+++ narwhals/_plan/arrow/expr.py
@@ -10,9 +10,10 @@ from narwhals._plan.arrow import functions as fn
 from narwhals._plan.arrow.series import ArrowSeries as Series
 from narwhals._plan.arrow.typing import ChunkedOrScalarAny, NativeScalar, StoresNativeT_co
 from narwhals._plan.compliant.column import ExprDispatch
+from narwhals._plan.compliant.expr import EagerExpr
+from narwhals._plan.compliant.scalar import EagerScalar
 from narwhals._plan.compliant.typing import namespace
 from narwhals._plan.expressions import NamedIR
-from narwhals._plan.protocols import EagerExpr, EagerScalar
 from narwhals._utils import (
     Implementation,
     Version,
