+++ narwhals/_plan/compliant/scalar.py
@@ -1,1 +1,131 @@
 from __future__ import annotations
+
+from typing import TYPE_CHECKING, Any, Protocol
+
+from narwhals._plan.compliant.expr import CompliantExpr, EagerExpr, LazyExpr
+from narwhals._plan.compliant.typing import FrameT_contra, LengthT, SeriesT, SeriesT_co
+
+if TYPE_CHECKING:
+    from typing_extensions import Self
+
+    from narwhals._plan import expressions as ir
+    from narwhals._plan.expressions import aggregation as agg
+    from narwhals._utils import Version
+    from narwhals.typing import IntoDType, PythonLiteral
+
+
+class CompliantScalar(
+    CompliantExpr[FrameT_contra, SeriesT_co], Protocol[FrameT_contra, SeriesT_co]
+):
+    _name: str
+
+    @property
+    def name(self) -> str:
+        return self._name
+
+    @classmethod
+    def from_python(
+        cls,
+        value: PythonLiteral,
+        name: str = "literal",
+        /,
+        *,
+        dtype: IntoDType | None,
+        version: Version,
+    ) -> Self: ...
+    def _with_evaluated(self, evaluated: Any, name: str) -> Self:
+        """Expr is based on a series having these via accessors, but a scalar needs to keep passing through."""
+        cls = type(self)
+        obj = cls.__new__(cls)
+        obj._evaluated = evaluated
+        obj._name = name or self.name
+        obj._version = self.version
+        return obj
+
+    def max(self, node: agg.Max, frame: FrameT_contra, name: str) -> Self:
+        """Returns self."""
+        return self._with_evaluated(self._evaluated, name)
+
+    def min(self, node: agg.Min, frame: FrameT_contra, name: str) -> Self:
+        """Returns self."""
+        return self._with_evaluated(self._evaluated, name)
+
+    def sum(self, node: agg.Sum, frame: FrameT_contra, name: str) -> Self:
+        """Returns self."""
+        return self._with_evaluated(self._evaluated, name)
+
+    def first(self, node: agg.First, frame: FrameT_contra, name: str) -> Self:
+        """Returns self."""
+        return self._with_evaluated(self._evaluated, name)
+
+    def last(self, node: agg.Last, frame: FrameT_contra, name: str) -> Self:
+        """Returns self."""
+        return self._with_evaluated(self._evaluated, name)
+
+    def _cast_float(self, node: ir.ExprIR, frame: FrameT_contra, name: str) -> Self:
+        """`polars` interpolates a single scalar as a float."""
+        dtype = self.version.dtypes.Float64()
+        return self.cast(node.cast(dtype), frame, name)
+
+    def mean(self, node: agg.Mean, frame: FrameT_contra, name: str) -> Self:
+        return self._cast_float(node.expr, frame, name)
+
+    def median(self, node: agg.Median, frame: FrameT_contra, name: str) -> Self:
+        return self._cast_float(node.expr, frame, name)
+
+    def quantile(self, node: agg.Quantile, frame: FrameT_contra, name: str) -> Self:
+        return self._cast_float(node.expr, frame, name)
+
+    def n_unique(self, node: agg.NUnique, frame: FrameT_contra, name: str) -> Self:
+        """Returns 1."""
+        ...
+
+    def std(self, node: agg.Std, frame: FrameT_contra, name: str) -> Self:
+        """Returns null."""
+        ...
+
+    def var(self, node: agg.Var, frame: FrameT_contra, name: str) -> Self:
+        """Returns null."""
+        ...
+
+    def arg_min(self, node: agg.ArgMin, frame: FrameT_contra, name: str) -> Self:
+        """Returns 0."""
+        ...
+
+    def arg_max(self, node: agg.ArgMax, frame: FrameT_contra, name: str) -> Self:
+        """Returns 0."""
+        ...
+
+    def count(self, node: agg.Count, frame: FrameT_contra, name: str) -> Self:
+        """Returns 0 if null, else 1."""
+        ...
+
+    def len(self, node: agg.Len, frame: FrameT_contra, name: str) -> Self:
+        """Returns 1."""
+        ...
+
+    def sort(self, node: ir.Sort, frame: FrameT_contra, name: str) -> Self:
+        return self._with_evaluated(self._evaluated, name)
+
+    def sort_by(self, node: ir.SortBy, frame: FrameT_contra, name: str) -> Self:
+        return self._with_evaluated(self._evaluated, name)
+
+    # NOTE: `Filter` behaves the same, (maybe) no need to override
+
+
+class EagerScalar(
+    CompliantScalar[FrameT_contra, SeriesT],
+    EagerExpr[FrameT_contra, SeriesT],
+    Protocol[FrameT_contra, SeriesT],
+):
+    def __len__(self) -> int:
+        return 1
+
+    def to_python(self) -> PythonLiteral: ...
+
+
+class LazyScalar(
+    CompliantScalar[FrameT_contra, SeriesT],
+    LazyExpr[FrameT_contra, SeriesT, LengthT],
+    Protocol[FrameT_contra, SeriesT, LengthT],
+): ...
