--- narwhals/_plan/protocols.py
@@ -1,269 +1,0 @@
-"""TODO: Split this module up into `narwhals._plan.compliant.*`."""
-
-from __future__ import annotations
-
-from typing import TYPE_CHECKING, Any, Protocol
-
-from narwhals._plan.compliant.column import EagerBroadcast, SupportsBroadcast
-from narwhals._plan.compliant.typing import (
-    FrameT_contra,
-    LengthT,
-    SeriesT,
-    SeriesT_co,
-    StoresVersion,
-)
-from narwhals._utils import Version
-
-if TYPE_CHECKING:
-    from typing_extensions import Self
-
-    from narwhals._plan import expressions as ir
-    from narwhals._plan.expressions import (
-        BinaryExpr,
-        FunctionExpr,
-        aggregation as agg,
-        boolean,
-        functions as F,
-    )
-    from narwhals._plan.expressions.boolean import IsBetween, IsFinite, IsNan, IsNull, Not
-    from narwhals.typing import IntoDType, PythonLiteral
-
-
-class CompliantExpr(StoresVersion, Protocol[FrameT_contra, SeriesT_co]):
-    """Everything common to `Expr`/`Series` and `Scalar` literal values."""
-
-    _evaluated: Any
-    """Compliant or native value."""
-
-    @property
-    def name(self) -> str: ...
-    @classmethod
-    def from_native(
-        cls, native: Any, name: str = "", /, version: Version = Version.MAIN
-    ) -> Self: ...
-    def _with_native(self, native: Any, name: str, /) -> Self:
-        return self.from_native(native, name or self.name, self.version)
-
-    # series & scalar
-    def abs(self, node: FunctionExpr[F.Abs], frame: FrameT_contra, name: str) -> Self: ...
-    def cast(self, node: ir.Cast, frame: FrameT_contra, name: str) -> Self: ...
-    def pow(self, node: FunctionExpr[F.Pow], frame: FrameT_contra, name: str) -> Self: ...
-    def not_(self, node: FunctionExpr[Not], frame: FrameT_contra, name: str) -> Self: ...
-    def fill_null(
-        self, node: FunctionExpr[F.FillNull], frame: FrameT_contra, name: str
-    ) -> Self: ...
-    def is_between(
-        self, node: FunctionExpr[IsBetween], frame: FrameT_contra, name: str
-    ) -> Self: ...
-    def is_finite(
-        self, node: FunctionExpr[IsFinite], frame: FrameT_contra, name: str
-    ) -> Self: ...
-    def is_nan(
-        self, node: FunctionExpr[IsNan], frame: FrameT_contra, name: str
-    ) -> Self: ...
-    def is_null(
-        self, node: FunctionExpr[IsNull], frame: FrameT_contra, name: str
-    ) -> Self: ...
-    def binary_expr(self, node: BinaryExpr, frame: FrameT_contra, name: str) -> Self: ...
-    def ternary_expr(
-        self, node: ir.TernaryExpr, frame: FrameT_contra, name: str
-    ) -> Self: ...
-    def over(self, node: ir.WindowExpr, frame: FrameT_contra, name: str) -> Self: ...
-    # NOTE: `Scalar` is returned **only** for un-partitioned `OrderableAggExpr`
-    # e.g. `nw.col("a").first().over(order_by="b")`
-    def over_ordered(
-        self, node: ir.OrderedWindowExpr, frame: FrameT_contra, name: str
-    ) -> Self | CompliantScalar[FrameT_contra, SeriesT_co]: ...
-    def map_batches(
-        self, node: ir.AnonymousExpr, frame: FrameT_contra, name: str
-    ) -> Self: ...
-    def rolling_expr(
-        self, node: ir.RollingExpr, frame: FrameT_contra, name: str
-    ) -> Self: ...
-    # series only (section 3)
-    def sort(self, node: ir.Sort, frame: FrameT_contra, name: str) -> Self: ...
-    def sort_by(self, node: ir.SortBy, frame: FrameT_contra, name: str) -> Self: ...
-    def filter(self, node: ir.Filter, frame: FrameT_contra, name: str) -> Self: ...
-    # series -> scalar
-    def first(
-        self, node: agg.First, frame: FrameT_contra, name: str
-    ) -> CompliantScalar[FrameT_contra, SeriesT_co]: ...
-    def last(
-        self, node: agg.Last, frame: FrameT_contra, name: str
-    ) -> CompliantScalar[FrameT_contra, SeriesT_co]: ...
-    def arg_min(
-        self, node: agg.ArgMin, frame: FrameT_contra, name: str
-    ) -> CompliantScalar[FrameT_contra, SeriesT_co]: ...
-    def arg_max(
-        self, node: agg.ArgMax, frame: FrameT_contra, name: str
-    ) -> CompliantScalar[FrameT_contra, SeriesT_co]: ...
-    def sum(
-        self, node: agg.Sum, frame: FrameT_contra, name: str
-    ) -> CompliantScalar[FrameT_contra, SeriesT_co]: ...
-    def n_unique(
-        self, node: agg.NUnique, frame: FrameT_contra, name: str
-    ) -> CompliantScalar[FrameT_contra, SeriesT_co]: ...
-    def std(
-        self, node: agg.Std, frame: FrameT_contra, name: str
-    ) -> CompliantScalar[FrameT_contra, SeriesT_co]: ...
-    def var(
-        self, node: agg.Var, frame: FrameT_contra, name: str
-    ) -> CompliantScalar[FrameT_contra, SeriesT_co]: ...
-    def quantile(
-        self, node: agg.Quantile, frame: FrameT_contra, name: str
-    ) -> CompliantScalar[FrameT_contra, SeriesT_co]: ...
-    def count(
-        self, node: agg.Count, frame: FrameT_contra, name: str
-    ) -> CompliantScalar[FrameT_contra, SeriesT_co]: ...
-    def len(
-        self, node: agg.Len, frame: FrameT_contra, name: str
-    ) -> CompliantScalar[FrameT_contra, SeriesT_co]: ...
-    def max(
-        self, node: agg.Max, frame: FrameT_contra, name: str
-    ) -> CompliantScalar[FrameT_contra, SeriesT_co]: ...
-    def mean(
-        self, node: agg.Mean, frame: FrameT_contra, name: str
-    ) -> CompliantScalar[FrameT_contra, SeriesT_co]: ...
-    def median(
-        self, node: agg.Median, frame: FrameT_contra, name: str
-    ) -> CompliantScalar[FrameT_contra, SeriesT_co]: ...
-    def min(
-        self, node: agg.Min, frame: FrameT_contra, name: str
-    ) -> CompliantScalar[FrameT_contra, SeriesT_co]: ...
-    def all(
-        self, node: FunctionExpr[boolean.All], frame: FrameT_contra, name: str
-    ) -> CompliantScalar[FrameT_contra, SeriesT_co]: ...
-    def any(
-        self, node: FunctionExpr[boolean.Any], frame: FrameT_contra, name: str
-    ) -> CompliantScalar[FrameT_contra, SeriesT_co]: ...
-
-
-class CompliantScalar(
-    CompliantExpr[FrameT_contra, SeriesT_co], Protocol[FrameT_contra, SeriesT_co]
-):
-    _name: str
-
-    @property
-    def name(self) -> str:
-        return self._name
-
-    @classmethod
-    def from_python(
-        cls,
-        value: PythonLiteral,
-        name: str = "literal",
-        /,
-        *,
-        dtype: IntoDType | None,
-        version: Version,
-    ) -> Self: ...
-    def _with_evaluated(self, evaluated: Any, name: str) -> Self:
-        """Expr is based on a series having these via accessors, but a scalar needs to keep passing through."""
-        cls = type(self)
-        obj = cls.__new__(cls)
-        obj._evaluated = evaluated
-        obj._name = name or self.name
-        obj._version = self.version
-        return obj
-
-    def max(self, node: agg.Max, frame: FrameT_contra, name: str) -> Self:
-        """Returns self."""
-        return self._with_evaluated(self._evaluated, name)
-
-    def min(self, node: agg.Min, frame: FrameT_contra, name: str) -> Self:
-        """Returns self."""
-        return self._with_evaluated(self._evaluated, name)
-
-    def sum(self, node: agg.Sum, frame: FrameT_contra, name: str) -> Self:
-        """Returns self."""
-        return self._with_evaluated(self._evaluated, name)
-
-    def first(self, node: agg.First, frame: FrameT_contra, name: str) -> Self:
-        """Returns self."""
-        return self._with_evaluated(self._evaluated, name)
-
-    def last(self, node: agg.Last, frame: FrameT_contra, name: str) -> Self:
-        """Returns self."""
-        return self._with_evaluated(self._evaluated, name)
-
-    def _cast_float(self, node: ir.ExprIR, frame: FrameT_contra, name: str) -> Self:
-        """`polars` interpolates a single scalar as a float."""
-        dtype = self.version.dtypes.Float64()
-        return self.cast(node.cast(dtype), frame, name)
-
-    def mean(self, node: agg.Mean, frame: FrameT_contra, name: str) -> Self:
-        return self._cast_float(node.expr, frame, name)
-
-    def median(self, node: agg.Median, frame: FrameT_contra, name: str) -> Self:
-        return self._cast_float(node.expr, frame, name)
-
-    def quantile(self, node: agg.Quantile, frame: FrameT_contra, name: str) -> Self:
-        return self._cast_float(node.expr, frame, name)
-
-    def n_unique(self, node: agg.NUnique, frame: FrameT_contra, name: str) -> Self:
-        """Returns 1."""
-        ...
-
-    def std(self, node: agg.Std, frame: FrameT_contra, name: str) -> Self:
-        """Returns null."""
-        ...
-
-    def var(self, node: agg.Var, frame: FrameT_contra, name: str) -> Self:
-        """Returns null."""
-        ...
-
-    def arg_min(self, node: agg.ArgMin, frame: FrameT_contra, name: str) -> Self:
-        """Returns 0."""
-        ...
-
-    def arg_max(self, node: agg.ArgMax, frame: FrameT_contra, name: str) -> Self:
-        """Returns 0."""
-        ...
-
-    def count(self, node: agg.Count, frame: FrameT_contra, name: str) -> Self:
-        """Returns 0 if null, else 1."""
-        ...
-
-    def len(self, node: agg.Len, frame: FrameT_contra, name: str) -> Self:
-        """Returns 1."""
-        ...
-
-    def sort(self, node: ir.Sort, frame: FrameT_contra, name: str) -> Self:
-        return self._with_evaluated(self._evaluated, name)
-
-    def sort_by(self, node: ir.SortBy, frame: FrameT_contra, name: str) -> Self:
-        return self._with_evaluated(self._evaluated, name)
-
-    # NOTE: `Filter` behaves the same, (maybe) no need to override
-
-
-class EagerExpr(
-    EagerBroadcast[SeriesT],
-    CompliantExpr[FrameT_contra, SeriesT],
-    Protocol[FrameT_contra, SeriesT],
-): ...
-
-
-class LazyExpr(
-    SupportsBroadcast[SeriesT, LengthT],
-    CompliantExpr[FrameT_contra, SeriesT],
-    Protocol[FrameT_contra, SeriesT, LengthT],
-): ...
-
-
-class EagerScalar(
-    CompliantScalar[FrameT_contra, SeriesT],
-    EagerExpr[FrameT_contra, SeriesT],
-    Protocol[FrameT_contra, SeriesT],
-):
-    def __len__(self) -> int:
-        return 1
-
-    def to_python(self) -> PythonLiteral: ...
-
-
-class LazyScalar(
-    CompliantScalar[FrameT_contra, SeriesT],
-    LazyExpr[FrameT_contra, SeriesT, LengthT],
-    Protocol[FrameT_contra, SeriesT, LengthT],
-): ...
